refactor herminebot_behaviors sources
