+++ robot/ros_ws/src/herminebot_behaviors/src/navigator/navigator_preemption.cpp
@@ -1,4 +1,4 @@
-#include "herminebot_behaviors/navigator_preemption.hpp"
+#include "herminebot_behaviors/navigator/navigator_preemption.hpp"
 
 namespace hrc_bt_navigator
 {
